aws-ec2-arm64: allow to use runqemu
The AWS-EC2 machines does not inherit from the class qemuboot[1] as the
QEMU machines[1] does. As a consequence, they cannot use the script
runqemu directly.

This adds the required bits to allow the AWS-EC2 machine (arm64 only) to
exec runqemu[2].

Note: The machines INHERIT the class qemuboot[3] and EXTRA_IMAGEDEPENDS
on qemu-system-native and its task do_addto_recipe_sysroot[4]. Also, the
machines sets the minimal subset of the appropriate qemuboot variables
related to its architecture arm64[5] (i.e. QB_SYSTEM_NAME, QB_CPU,
QB_CPU_KVM...).

Note 2: The recipe edk2-firmware requires to set the following variables
to be build against the AWS-EC2 ARM64 machine:
 - COMPATIBLE_MACHINE
 - EDK2_PLATFORM
 - EDK2_PLATFORM_DSC
 - EDK2_BIN_NAME

Note 3: The UEFI firmware for aws-ec2-arm64 is installed as uefi.bin[6]
unlike the machine machine qemuarm64 that installs the firmware as
${EDK2_BIN_NAME}[7] aka QEMU_EFI.fd[8].

Note 4: The QEMU CPU is set to neoverse-n1 as meta-aws tunes the machine
aws-ec2-arm64 to armv8.2a since commit[9]. It is the ISA used since by
Graviton2, according to issue[10] and Wikipedia[11]:

	Graviton2

	The Graviton2 CPU has 64 Neoverse N1 cores, with ARMv8.2-A ISA
	including 2×128 bit Neon, LSE, fp16, rcpc, dotprod, crypto. The
	vCPUs are physical cores in a single NUMA domain, running at 2.5
	GHz.

	EC2 instances with Graviton2 CPU: M6g, M6gd, C6g, C6gd, C6gn,
	R6g, R6gd, T4g, X2gd, G5g, Im4gn, Is4gen, I4g. One or more of
	these instances are available in 28 AWS regions.

[1]: https://git.yoctoproject.org/poky/tree/meta/classes-recipe/qemuboot.bbclass
[2]: https://docs.yoctoproject.org/dev-manual/qemu.html
[3]: https://git.yoctoproject.org/poky/tree/meta/conf/machine/include/qemu.inc#n30
[4]: https://git.yoctoproject.org/poky/tree/meta/classes-recipe/qemuboot.bbclass#n190
[5]: https://git.yoctoproject.org/poky/tree/meta/conf/machine/qemuarm64.conf
[6]: https://git.yoctoproject.org/meta-arm/tree/meta-arm/recipes-bsp/uefi/edk2-firmware.inc#n123
[7]: https://git.yoctoproject.org/meta-arm/tree/meta-arm/recipes-bsp/uefi/edk2-firmware_%25.bbappend#n17
[8]: https://git.yoctoproject.org/meta-arm/tree/meta-arm/recipes-bsp/uefi/edk2-firmware_%25.bbappend#n9
[9]: https://github.com/aws4embeddedlinux/meta-aws/commit/5a69468854b0519aed940546ef9815ece3c7b6a2
[10]: https://github.com/aws4embeddedlinux/meta-aws/issues/6534
[11]: https://en.wikipedia.org/wiki/AWS_Graviton