qemu{x86-64,arm64}: allow to use UEFI
This allows to boot with UEFI using runqem[1] and OVMF[2] (x86-64) or
EDK II[3] (ARM64).

Note: It adds the OVMF or EDK II firmware as an EXTRA_IMAGEDEPENDS[4],
IMAGE_INSTALL[5]'s grub-efi and sets it as the EFI_PROVIDER[6], and uses
the WKS_FILE genericx86.wks.in from meta-yocto-bsp[7] (x86-64) or
WKS_FILE efi-disk.wks.in from meta-arm[8] (ARM64).

[1]: https://docs.yoctoproject.org/dev-manual/qemu.html
[2]: https://github.com/tianocore/tianocore.github.io/wiki/OVMF
[3]: https://github.com/tianocore/tianocore.github.io/wiki/edk-ii
[4]: https://docs.yoctoproject.org/ref-manual/variables.html#term-EXTRA_IMAGEDEPENDS
[5]: https://docs.yoctoproject.org/ref-manual/variables.html#term-IMAGE_INSTALL
[6]: https://docs.yoctoproject.org/ref-manual/variables.html#term-EFI_PROVIDER
[7]: https://git.yoctoproject.org/poky/tree/meta-yocto-bsp/wic/genericx86.wks.in
[8]: https://git.yoctoproject.org/meta-arm/tree/meta-arm/wic/efi-disk.wks.in